code/requirements.txt
-f file:///datastorage02/share/packages/python
backports.functools-lru-cache==1.4
cycler==0.10.0
deap==1.2.2
matplotlib==2.1.0
numpy==1.13.3
pandas==0.22.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7
tqdm==4.19.5
